Refactor: Tipos das propriedades relacionadas a armazenamento (float -> int)

+++ Models/Aplicativo.cs
@@ -3,10 +3,10 @@ namespace Models;
 public class Aplicativo {
     // Declaração de propriedades
     public string Nome { get; set; }
-    public float Tamanho { get; set; }
+    public int Tamanho { get; set; }
 
     // Construção da classe
-    public Aplicativo(String nome, float tamanho) {
+    public Aplicativo(String nome, int tamanho) {
         Nome = nome;
         Tamanho = tamanho;
     }
